♻️ Migrate the unit tests about object *GitHubOperationRunner* into correct test module and move common test value into another single module.
