+++ github-label-management/test/unit_test/_values.py
@@ -1,0 +1,12 @@
+SAMPLE_YAML = """
+repositories:
+  - my-org/my-repository
+labels:
+  Bug:
+    color: d73a4a
+    description: Something went wrong.
+  Enhancement:
+    color: 005cc5
+    description: New feature or improvement.
+delete_unused: true
+"""
